DABOM-229 refactor: port 디렉토리 제거
