+++ src/main/java/com/project/domain/usage/infra/messaging/UsageRealtimeEventPublisher.java
@@ -1,4 +1,4 @@
-package com.project.domain.usage.service.port;
+package com.project.domain.usage.infra.messaging;
 
 import com.project.global.event.dto.usage.UsageRealtimePayload;
