+++ src/main/java/com/project/domain/usage/infra/messaging/UsageRealtimeKafkaProducer.java
@@ -3,7 +3,6 @@ package com.project.domain.usage.infra.messaging;
 import org.springframework.kafka.core.KafkaTemplate;
 import org.springframework.stereotype.Component;
 
-import com.project.domain.usage.service.port.UsageRealtimeEventPublisher;
 import com.project.global.event.dto.EventEnvelope;
 import com.project.global.event.dto.usage.UsageRealtimePayload;
